Give the item grid panels one implementation, starting with the bank
The bank and inventory panels were separate implementations of the same
panel: 87% identical after allowing for their vocabulary, sharing
twenty-four identically named private methods. That is not a tidiness
problem. It is why the capacity readout counted icons instead of items in
the bank for as long as it did -- the same defect was found and fixed in
the inventory, and the bank went on being wrong until it was reported
again from play. A second copy does not double the work of a fix; it
halves the chance of one.

UITKItemGridPanel now holds the grid: slot construction, visuals, the
capacity readout, drag and drop, tooltips, the operation tracker, and the
container subscriptions. UITKBank keeps what is actually particular to a
bank and is 69 lines instead of 986.

What made this straightforward is that the domain model had already
unified them. IBankController and IInventoryController both derive from
IItemContainer, which exposes everything the view uses -- the slot list,
the emptiness test, the item accessor and the two change events. The
panels simply were not using it, so the base needs no generics and knows
nothing about which container it draws.

Four abstract members carry every difference: the element and USS name
prefix, the drag identity, the container's name in a request, and the
swap broadcast. The ten panel-specific names all followed {prefix}-suffix
without exception, so one string replaces twenty constants.

The inventory is deliberately not migrated in this commit. These panels
have no behavioural tests -- nothing covers drag, drop, slot rendering or
lock overlays -- so moving both at once would put the whole item UI on one
untested change. The bank alone can be exercised and confirmed first.

Verified: no serialized fields existed on the panel, the script GUID and
class identity are unchanged so the scene binding holds, every method of
the old panel is accounted for in the new pair, and the two renamed public
handlers had no callers outside the panel.

Full EditMode suite: 1260 tests, 1260 passed.
